refactor: move timeline color setting handling to main
This way we only need vscode host service in main.
